Migrate merge test
