+++ test/unit.html
@@ -36,7 +36,6 @@
   <script src="hasOwn_test_old.js"></script>
   <script src="isArray_test_old.js"></script>
   <script src="isPlainObject_test_old.js"></script>
-  <script src="merge_test_old.js"></script>
   <script src="processCommand_test_old.js"></script>
   <script src="processStates_test_old.js"></script>
 </body>
